# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hive",
    "version": "0.0.1",
    "description": "Hive API"
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/auth/exists": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Check if a user exists",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "minLength": 3,
                    "type": "string"
                  }
                },
                "required": [
                  "username"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "exists"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Verify the user is authenticated",
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "User is authenticated"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Logout the user",
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get the user for the current request",
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "name"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "description": "Update a user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newPassword": {
                    "minLength": 8,
                    "type": "string"
                  },
                  "oldPassword": {
                    "minLength": 8,
                    "type": "string"
                  }
                },
                "required": [
                  "newPassword",
                  "oldPassword"
                ]
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "204": {
            "description": "User updated successfully"
          },
          "401": {
            "description": "Invalid username or password",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid username or password",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/channels": {
      "get": {
        "tags": [
          "Channels"
        ],
        "description": "Get a list of channels",
        "parameters": [
          {
            "schema": {
              "default": 1,
              "type": "number"
            },
            "in": "query",
            "name": "page",
            "required": true
          }
        ],
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "channels": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "customUrl": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "description",
                          "tags",
                          "customUrl",
                          "name"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "channels",
                    "total"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Channels"
        ],
        "description": "Add a channel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "201": {
            "description": "Channel added successfully"
          },
          "404": {
            "description": "Channel does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Channel does not exist",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Channel already downloaded",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Channel already downloaded",
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/channels/{channelId}": {
      "get": {
        "tags": [
          "Channels"
        ],
        "description": "Get a channel",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "channelId",
            "required": true
          }
        ],
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "customUrl": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "description",
                    "tags",
                    "customUrl",
                    "name"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Channel not found"
          }
        }
      }
    },
    "/channels/{channelId}/scan": {
      "post": {
        "tags": [
          "Channels"
        ],
        "description": "Scan a channel for new metadata and assets",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "channelId",
            "required": true
          }
        ],
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "201": {
            "description": "Scan started successfully"
          }
        }
      }
    },
    "/channels/{channelId}/stats": {
      "get": {
        "tags": [
          "Channels"
        ],
        "description": "Get stats about a channel",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "channelId",
            "required": true
          }
        ],
        "security": [
          {
            "apikey": [
              "x-api-key"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "videos": {
                      "type": "number"
                    },
                    "streams": {
                      "type": "number"
                    },
                    "shorts": {
                      "type": "number"
                    },
                    "directorySize": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "videos",
                    "streams",
                    "shorts",
                    "directorySize"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Channel not found"
          }
        }
      }
    },
...
  },
...
}